refactor(rate):rename files
