+++ src/app/live-demo/rate/rate/rate-full.component.html
