+++ src/app/live-demo/rate/rate/rate-full.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-app',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  selector: 'app-rate-full',
+  templateUrl: './rate-full.component.html',
+  styleUrls: ['./rate-full.component.scss']
 })
 export class RateFullComponent implements OnInit {
